Extract Date and Entry
I'm not sure what idiomatic Elm looks like so I'm basing this mostly on
https://github.com/rtfeldman/elm-spa-example

Instead of Date and Entry being type aliases for records with a specific
structure they are custom types with a single variant, which holds the
actual record. This prevents direct access to the data from outside the
model. I like that, but there is a lot of biolerplate...